installing firebase to app
