+++ index.html
@@ -1,30 +1,31 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover" />
-  <meta name="Description" content="Put your description here.">
-  <base href="/">
-
-  <style>
-    html,
-    body {
-      margin: 0;
-      padding: 0;
-      font-family: sans-serif;
-      background-color: #ededed;
-    }
-  </style>
-  <title>video-courses</title>
-</head>
-
-<body>
-  <a href="/home">home</a>
-  
-  <div id="content"></div>
-
-  <script type="module" src="./out-tsc/src/scripts/router/index.js"></script>
-</body>
-
+  <head>
+    <meta charset="utf-8" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1.0, viewport-fit=cover"
+    />
+    <meta name="Description" content="Put your description here." />
+    <base href="/" />
+    <script type="module" src="./out-tsc/src/scripts/firebase/firebase.js"></script>
+    <style>
+      html,
+      body {
+        margin: 0;
+        padding: 0;
+        font-family: sans-serif;
+        background-color: #ededed;
+      }
+      </style>
+    <title>video-courses</title>
+    </head>
+    
+    <body>
+      <a href="/home">home</a>
+      
+      <div id="content"></div>
+      
+    <script type="module" src="./out-tsc/src/scripts/router/index.js"></script>
+  </body>
 </html>
