--- src/scripts/auth/firebase.js
@@ -1,18 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyD15uPmbdoJpgOL4nKJ8j7GFakRDwUT8W4",
-  authDomain: "stable-glass-258919.firebaseapp.com",
-  databaseURL: "https://stable-glass-258919.firebaseio.com",
-  projectId: "stable-glass-258919",
-  storageBucket: "stable-glass-258919.appspot.com",
-  messagingSenderId: "103863675984",
-  appId: "1:103863675984:web:13ccfb8dca66b89a7f7555"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
