+++ src/scripts/router/index.ts
@@ -2,8 +2,6 @@ import { Route, Router } from '@vaadin/router'; // for Webpack / Polymer CLI
 import '../../views/video-courses.js';
 import '../../views/protected/Home.js';
 
-
-
 const routes: Route[] = [
   { path: '/', component: 'video-courses' },
   { path: '/home', component: 'vid-home' },
