+++ src/views/VideoCourses.ts
@@ -52,28 +52,20 @@ export class VideoCourses extends LitElement {
   render() {
     return html`
       <main>
+      
         <div class="logo"><img alt="open-wc logo" src=${logo} /></div>
         <h1>${this.title}</h1>
-
+      
         <p>Edit <code>src/VideoCourses.ts</code> and save to reload.</p>
-        <a
-          class="app-link"
-          href="https://open-wc.org/guides/developing-components/code-examples"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
+        <a class="app-link" href="https://open-wc.org/guides/developing-components/code-examples" target="_blank"
+          rel="noopener noreferrer">
           Code examples
         </a>
       </main>
-
+      
       <p class="app-footer">
         🚽 Made with love by
-        <a
-          target="_blank"
-          rel="noopener noreferrer"
-          href="https://github.com/open-wc"
-          >open-wc</a
-        >.
+        <a target="_blank" rel="noopener noreferrer" href="https://github.com/open-wc">open-wc</a>.
       </p>
     `;
   }
